fix: isDeleted() 도메인 책임 부여 및 조회 메서드 용도에 따라 분리

+++ src/main/java/roomescape/service/ReservationTimeCommandService.java
@@ -7,7 +7,9 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.exception.ErrorMessage;
+import roomescape.exception.custom.BadRequestException;
 import roomescape.exception.custom.ConflictException;
+import roomescape.exception.custom.NotFoundException;
 import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
 
@@ -27,9 +29,17 @@ public class ReservationTimeCommandService {
 
     @Transactional
     public void delete(long reservationTimeId) {
+        ReservationTime time = reservationTimeDao.findByTimeId(reservationTimeId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.TIME_NOT_FOUND));
+
+        if (time.isDeleted()) {
+            throw new BadRequestException(ErrorMessage.TIME_ALREADY_DELETED);
+        }
+
         if (reservationDao.existsByTimeId(reservationTimeId)) {
             throw new ConflictException(ErrorMessage.TIME_IN_USE);
         }
+
         reservationTimeDao.deleteByTimeId(reservationTimeId);
     }
